TextCNN/requirements.txt
tqdm==4.54.1
numpy==1.19.4
torch==1.5.0
PyYAML==5.3.1
scikit_learn==0.24.0
